Touched up some of the api pathing
Bug: when you add new spools, the interests aren't inherited
This shouldn't be happening, but
it seems like the methods being used are not properly
setting interests for usersettings

+++ threadit-api/Repositories/InterestRepository.cs
@@ -49,7 +49,7 @@ namespace ThreaditAPI.Repositories {
             if(interest == null)
                 throw new Exception("Interest does not exist");
             int spoolCount = await DeIterateSpoolCount(interest);
-            if(spoolCount <= 0){
+            if(spoolCount <= 0 && interestName != "General"){
                 db.Interests.Remove(interest);
                 await db.SaveChangesAsync();
             }
